% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{classmates}
\alias{classmates}
\title{Starwars names}
\format{
character vector
}
\source{
The mind of George Lucas \url{https://www.starwars.com}
}
\usage{
classmates
}
\description{
A vector of 10 randomly chosen names from the Starwars universe.
}
\keyword{datasets}
