% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_couple_initials.R
\name{make_couple_initials}
\alias{make_couple_initials}
\title{Creating a combination of initials for two names}
\usage{
make_couple_initials(first_name, second_name)
}
\arguments{
\item{first_name}{a string, its initials will be the part before "_"}

\item{second_name}{a string, its initials will be the part after "_"}
}
\value{
a string of initials separated by "_"
}
\description{
make_couple_initials(s1, s2) returns a string of initials of s1 and s2
separated by "_"
}
