% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_groups.R
\name{make_groups}
\alias{make_groups}
\title{Make couples from a list of names}
\usage{
make_groups(names)
}
\arguments{
\item{names}{vector of names}
}
\value{
data frame of coupled names
}
\description{
make_groups creates a data frame of couples from the given set of names
}
