% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_groups_by_time.R
\name{make_groups_by_time}
\alias{make_groups_by_time}
\title{Assign meeting time for people}
\usage{
make_groups_by_time(names)
}
\arguments{
\item{names}{a character vector, to make couples}
}
\value{
a data frame with couples (person1, person2) and meeting time
}
\description{
make_groups_by_time(names) makes couples with meeting time for a given set
of names
}
