% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_initials.R
\name{return_initials}
\alias{return_initials}
\title{Extract initials for a given string}
\usage{
return_initials(person_name)
}
\arguments{
\item{person_name}{string, initials will be extracted}
}
\value{
the initials of the string
}
\description{
return_initials(s) returns the initials of a string s. If the string is
separated by space, multiple letters will be returned.
}
