% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{starwars_coffee_people}
\alias{starwars_coffee_people}
\title{Star Wars names}
\format{
character vector
}
\source{
Star Wars Universe \url{https://www.starwars.com}
}
\usage{
starwars_coffee_people
}
\description{
A vector of 40 randomly chosen names from Star Wars Universe
}
\keyword{datasets}
