% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_groups.R
\name{make_groups}
\alias{make_groups}
\title{Group people in pairs}
\usage{
make_groups(names, result = T)
}
\arguments{
\item{names}{(character) vector with names of participants}

\item{result}{Boolean flag, indicating whether a result should be returned (default = True)}
}
\value{
matrix of coupled names
}
\description{
Turn a vector of names into a matrix with randomly
paired people, preparing for a mystery coffee
date!
}
\examples{
make_groups(c("Jacques","Georgi"))

}
